Goal-Guru/requirements.txt
flask>=3.1.0
flask-cors>=6.0.0
flask-caching>=2.3.0
flask-socketio>=5.5.0
flask-sqlalchemy>=3.1.1
flask-limiter>=3.12
flask-swagger-ui>=5.21.0
requests>=2.32.0
pytz>=2025.1
numpy<2.0.0
pandas>=2.2.3
scikit-learn>=1.6.1
scipy>=1.15.2
psycopg2-binary>=2.9.10
python-socketio>=5.12.0
joblib>=1.4.2
werkzeug>=3.1.0
python-dateutil>=2.9.0
glicko2>=2.1.0
trueskill>=0.4.5
xgboost>=3.0.0
sklearn-crfsuite>=0.5.0
python-crfsuite>=0.9.11
tqdm>=4.67.1
aiohttp>=3.12.14
python-dotenv>=1.1.1
eventlet>=0.40.2
geopy>=2.4.1
email-validator>=2.2.0
tabulate>=0.9.0
